src/main/webapp/app/entities/delivery-address/route/delivery-address-routing-resolve.service.spec.ts
import {TestBed} from '@angular/core/testing';
import {HttpResponse} from '@angular/common/http';
import {HttpClientTestingModule} from '@angular/common/http/testing';
import {ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap, RouterStateSnapshot} from '@angular/router';
import {RouterTestingModule} from '@angular/router/testing';
import {of} from 'rxjs';

import {IDeliveryAddress} from '../delivery-address.model';
import {DeliveryAddressService} from '../service/delivery-address.service';

import deliveryAddressResolve from './delivery-address-routing-resolve.service';

describe('DeliveryAddress routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let service: DeliveryAddressService;
  let resultDeliveryAddress: IDeliveryAddress | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    service = TestBed.inject(DeliveryAddressService);
    resultDeliveryAddress = undefined;
  });

  describe('resolve', () => {
    it('should return IDeliveryAddress returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({body: {id}})));
      mockActivatedRouteSnapshot.params = {id: 123};

      // WHEN
      TestBed.runInInjectionContext(() => {
        deliveryAddressResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultDeliveryAddress = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDeliveryAddress).toEqual({id: 123});
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      TestBed.runInInjectionContext(() => {
        deliveryAddressResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultDeliveryAddress = result;
          },
        });
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultDeliveryAddress).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IDeliveryAddress>({body: null})));
      mockActivatedRouteSnapshot.params = {id: 123};

      // WHEN
      TestBed.runInInjectionContext(() => {
        deliveryAddressResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultDeliveryAddress = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDeliveryAddress).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
